add an example for  context
