+++ torchmix/components/__init__.py
@@ -1,4 +1,5 @@
 from ._token import Token
+from .attentions import SelfAttention, WindowAttention
 from .avg_pool import AvgPool
 from .channel_mixer import ChannelMixer
 from .containers.add import Add
@@ -10,9 +11,7 @@ from .extract import Extract
 from .patch_embed import PatchEmbed
 from .patch_merging import PatchMerging
 from .position_embed import PositionEmbed
-from .self_attention import SelfAttention
 from .token_mixer import TokenMixer
-from .window_attention import WindowAttention
 
 __all__ = [
     "Attach",
