+++ torchmix/components/attentions.py
@@ -1,3 +1,4 @@
+import math
 from typing import Tuple
 
 from einops import einsum, rearrange, unpack
@@ -20,9 +21,9 @@ class SelfAttention(MixModule):
 
     def __init__(
         self,
-        dim: int,
-        num_heads: int,
-        head_dim: int,
+        dim: int = 768,
+        num_heads: int = 8,
+        head_dim: int = 64,
     ):
         self.inner_dim = num_heads * head_dim
         self.scale: float = head_dim**-0.5
@@ -50,7 +51,11 @@ class SelfAttention(MixModule):
 
     def to_qkv(
         self, x: Float[Tensor, "... d"]
-    ) -> Tuple[Float[Tensor, "... d"], Float[Tensor, "... d"], Float[Tensor, "... d"],]:
+    ) -> Tuple[
+        Float[Tensor, "... d"],
+        Float[Tensor, "... d"],
+        Float[Tensor, "... d"],
+    ]:
         query, key, value = unpack(
             self._to_qkv(x),
             [[self.inner_dim], [self.inner_dim], [self.inner_dim]],
@@ -102,3 +107,66 @@ class SelfAttention(MixModule):
         out = self.collect_heads(out)
         out = self._proj(out)
         return out
+
+
+class WindowAttention(SelfAttention):
+    """Local window attention layer from Swin-transformer
+
+    Example:
+        >>> model = WindowAttention(96, 8, 64)
+        >>> inputs = torch.randn(32, 56*56, 96)
+        >>> model(inputs).shape
+        torch.Size([32, 3136, 96])
+    """
+
+    def __init__(
+        self,
+        dim: int = 96,
+        window_size: int = 8,
+        num_heads: int = 8,
+        head_dim: int = 64,
+    ):
+        super().__init__(
+            dim=dim,
+            num_heads=num_heads,
+            head_dim=head_dim,
+        )
+        self.window_size: int = window_size
+
+    def split_qkv(
+        self,
+        query: Float[Tensor, "b n d_in"],
+        key: Float[Tensor, "b n d_in"],
+        value: Float[Tensor, "b n d_in"],
+    ) -> Tuple[
+        Float[Tensor, "b h w head window d_out"],
+        Float[Tensor, "b h w head window d_out"],
+        Float[Tensor, "b h w head window d_out"],
+    ]:
+        _batch_size, _seq_length, _inner_dim = query.shape
+        patch_size = round(math.sqrt(_seq_length))
+
+        query, key, value = map(
+            lambda x: rearrange(
+                x,
+                "... (h ph w pw) (hd d) -> ... h w hd (ph pw) d",
+                ph=self.window_size,
+                pw=self.window_size,
+                h=patch_size // self.window_size,
+                w=patch_size // self.window_size,
+                hd=self.num_heads,
+            ),
+            (query, key, value),
+        )
+        return query, key, value
+
+    def collect_heads(
+        self,
+        out: Float[Tensor, "... h w head window d_out"],
+    ) -> Float[Tensor, "... n d_in"]:
+        return rearrange(
+            out,
+            "... h w hd (ph pw) d -> ... (h ph w pw) (hd d)",
+            ph=self.window_size,
+            pw=self.window_size,
+        )
